feat:code is divided into modules. Added files data.js, util.js
